% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{build_report_markdown}
\alias{build_report_markdown}
\title{Assemble the full pandoc-markdown document for one format}
\usage{
build_report_markdown(
  report,
  format,
  assets_dir,
  orientation = "portrait",
  margin = "1in",
  title_page_yaml = NULL
)
}
\description{
Assemble the full pandoc-markdown document for one format
}
\keyword{internal}
