% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{commit_snapshot_metadata}
\alias{commit_snapshot_metadata}
\title{Commit a snapshot's metadata (manifest + ledger) and annotate a tag}
\usage{
commit_snapshot_metadata(root, meta_files, id, tag = NULL, verbose = 1L)
}
\description{
Commit a snapshot's metadata (manifest + ledger) and annotate a tag
}
\keyword{internal}
