% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{default_run_name}
\alias{default_run_name}
\title{Default run name: the running script's file name, else \code{interactive-session}}
\usage{
default_run_name()
}
\description{
Default run name: the running script's file name, else \code{interactive-session}
}
\keyword{internal}
