% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{grid_border_line}
\alias{grid_border_line}
\title{A grid-table border line (\verb{+---+}); alignment colons on the \code{=} header line}
\usage{
grid_border_line(widths, char, aligns = NULL)
}
\description{
A grid-table border line (\verb{+---+}); alignment colons on the \code{=} header line
}
\keyword{internal}
