% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{grid_content_line}
\alias{grid_content_line}
\title{A grid-table content line, supporting cells that span several columns}
\usage{
grid_content_line(segments, widths)
}
\description{
A grid-table content line, supporting cells that span several columns
}
\keyword{internal}
