% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{indent_depth}
\alias{indent_depth}
\title{Indent depth of table labels}
\usage{
indent_depth(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An integer vector.
}
\description{
The number of \code{\link[=indent]{indent()}} levels at the start of each label.
}
\keyword{internal}
