% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{installed_version_or_na}
\alias{installed_version_or_na}
\title{Installed version of a package as a string, or NA if it is not installed}
\usage{
installed_version_or_na(package)
}
\description{
Installed version of a package as a string, or NA if it is not installed
}
\keyword{internal}
