% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso8601}
\alias{iso8601}
\title{Human-readable extended ISO 8601 timestamp (for manifests)}
\usage{
iso8601(time = utc_now(), tz = "UTC")
}
\arguments{
\item{time}{A time; rendered in \code{tz}.}

\item{tz}{Timezone; default UTC.}
}
\description{
Human-readable extended ISO 8601 timestamp (for manifests)
}
