% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{normalise_banner_rows}
\alias{normalise_banner_rows}
\title{Normalise the \code{banner_rows} argument into a data frame of label + after}
\usage{
normalise_banner_rows(banner_rows, n_rows)
}
\description{
Normalise the \code{banner_rows} argument into a data frame of label + after
}
\keyword{internal}
