% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{normalise_group_headers}
\alias{normalise_group_headers}
\title{Normalise the \code{group_headers} argument into a data frame of label + span}
\usage{
normalise_group_headers(group_headers, n_columns)
}
\description{
Normalise the \code{group_headers} argument into a data frame of label + span
}
\keyword{internal}
