% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uob_scales.R
\name{pal_uob}
\alias{pal_uob}
\title{UoB palette generator, ggsci style}
\usage{
pal_uob(palette = "categorical", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Name of a palette in \link{uob_palettes}.}

\item{alpha}{Opacity in (0, 1].}

\item{reverse}{Reverse the palette.}
}
\value{
A function \verb{function(n)} returning \code{n} hex colours.
}
\description{
Returns a function that takes the number of colours needed and gives that
many UoB colours. Asking for more colours than the palette holds
interpolates between them with \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}.
}
\examples{
pal_uob()(3)
pal_uob("deep", alpha = 0.7)(3)
}
