% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{program_version_line}
\alias{program_version_line}
\title{First line of a program's version output, or NA if it is not on the PATH}
\usage{
program_version_line(program, version_args = "--version")
}
\description{
First line of a program's version output, or NA if it is not on the PATH
}
\keyword{internal}
