% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{record_checkpoint}
\alias{record_checkpoint}
\title{Write a checkpoint stamp into the open run log}
\usage{
record_checkpoint(log, stamp)
}
\description{
Write a checkpoint stamp into the open run log
}
\keyword{internal}
