% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{redcap_metadata}
\alias{redcap_metadata}
\title{Fetch and parse the REDCap data dictionary (\code{content = "metadata"})}
\usage{
redcap_metadata(url, token)
}
\arguments{
\item{url}{REDCap API URL.}

\item{token}{Resolved API token.}
}
\value{
A data frame (the data dictionary).
}
\description{
Fetch and parse the REDCap data dictionary (\code{content = "metadata"})
}
