% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_figure_section}
\alias{render_figure_section}
\title{Render a figure section (saving a plot object to PNG if needed)}
\usage{
render_figure_section(fig, assets_dir, index)
}
\description{
Render a figure section (saving a plot object to PNG if needed)
}
\keyword{internal}
