% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{render_table_gridtable}
\alias{render_table_gridtable}
\title{Alias of \code{\link[=render_table_markdown]{render_table_markdown()}} naming the grid-table target explicitly}
\usage{
render_table_gridtable(x)
}
\arguments{
\item{x}{A \code{bctu_report_table}.}
}
\value{
A pandoc grid table string.
}
\description{
Alias of \code{\link[=render_table_markdown]{render_table_markdown()}} naming the grid-table target explicitly
}
