% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{resolve_log_location}
\alias{resolve_log_location}
\title{Resolve the log location (explicit, or \verb{<project dir>/log-output})}
\usage{
resolve_log_location(location)
}
\description{
Resolve the log location (explicit, or \verb{<project dir>/log-output})
}
\keyword{internal}
