% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{resolve_table_columns}
\alias{resolve_table_columns}
\title{Resolve the \code{columns}/\code{align} arguments into a tidy column specification}
\usage{
resolve_table_columns(data, columns, align)
}
\description{
Resolve the \code{columns}/\code{align} arguments into a tidy column specification
}
\keyword{internal}
