% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{retag_upper}
\alias{retag_upper}
\title{Re-tag tagged NAs to their uppercase form (SAS special missings)}
\usage{
retag_upper(tbl)
}
\description{
Re-tag tagged NAs to their uppercase form (SAS special missings)
}
\keyword{internal}
