% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uob_scales.R
\name{scale_color_uob}
\alias{scale_color_uob}
\alias{scale_colour_uob}
\alias{scale_fill_uob}
\title{UoB discrete ggplot2 scales}
\usage{
scale_color_uob(palette = "categorical", alpha = 1, reverse = FALSE, ...)

scale_colour_uob(palette = "categorical", alpha = 1, reverse = FALSE, ...)

scale_fill_uob(palette = "categorical", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a palette in \link{uob_palettes}.}

\item{alpha}{Opacity in (0, 1].}

\item{reverse}{Reverse the palette.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale.
}
\description{
UoB discrete ggplot2 scales
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(mtcars, ggplot2::aes(factor(cyl), mpg, fill = factor(cyl))) +
    ggplot2::geom_boxplot() +
    scale_fill_uob("deep")
}
}
