% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uob_scales.R
\name{scale_color_uob_c}
\alias{scale_color_uob_c}
\alias{scale_colour_uob_c}
\alias{scale_fill_uob_c}
\title{UoB continuous ggplot2 scales}
\usage{
scale_color_uob_c(palette = "mono_blue", alpha = 1, reverse = FALSE, ...)

scale_colour_uob_c(palette = "mono_blue", alpha = 1, reverse = FALSE, ...)

scale_fill_uob_c(palette = "mono_blue", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a palette in \link{uob_palettes}.}

\item{alpha}{Opacity in (0, 1].}

\item{reverse}{Reverse the palette.}

\item{...}{Passed to \code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn()}} or
\code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn()}}.}
}
\value{
A ggplot2 scale.
}
\description{
UoB continuous ggplot2 scales
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg, colour = hp)) +
    ggplot2::geom_point() +
    scale_color_uob_c()
}
}
