% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_postprocess.R
\name{set_theme_colours}
\alias{set_theme_colours}
\title{Replace entries of a theme's colour scheme}
\usage{
set_theme_colours(theme, colours)
}
\arguments{
\item{theme}{The theme1.xml text.}

\item{colours}{Named hex colours, see \code{\link[=repair_report_docx]{repair_report_docx()}}.}
}
\value{
The theme1.xml text with those entries replaced.
}
\description{
Replace entries of a theme's colour scheme
}
\keyword{internal}
