% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uob_scales.R
\name{show_uob_palettes}
\alias{show_uob_palettes}
\title{Preview UoB palettes}
\usage{
show_uob_palettes(palettes = names(uob_palettes), n = 10L)
}
\arguments{
\item{palettes}{Palettes to show. Default: all.}

\item{n}{Number of colours per row.}
}
\value{
Invisible \code{NULL}.
}
\description{
Preview UoB palettes
}
\examples{
show_uob_palettes(c("deep", "standard"))
}
