% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sql_discover_objects}
\alias{sql_discover_objects}
\title{Run a discovery query and return the object names in its first column}
\usage{
sql_discover_objects(con, views_query, include = NULL, verbose = 2L)
}
\arguments{
\item{con}{A DBI connection.}

\item{views_query}{Discovery SQL.}

\item{include}{Optional regex to subset the names.}

\item{verbose}{Verbosity.}
}
\value{
A character vector of object names.
}
\description{
Run a discovery query and return the object names in its first column
}
