% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_theme.R
\name{tint}
\alias{tint}
\title{Lighten a hex colour towards white}
\usage{
tint(hex, amount)
}
\arguments{
\item{hex}{A six-digit hex colour.}

\item{amount}{The share of the way to white, 0 to 1.}
}
\value{
A six-digit hex colour.
}
\description{
Lighten a hex colour towards white
}
\keyword{internal}
