% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tool_version}
\alias{tool_version}
\title{First line of a command-line tool's version output}
\usage{
tool_version(cmd, flag)
}
\description{
First line of a command-line tool's version output
}
\keyword{internal}
