% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{list_snapshots}
\alias{list_snapshots}
\title{List snapshot ids in a store (newest last)}
\usage{
list_snapshots(store = snapshot_location(verbose = 0L))
}
\arguments{
\item{store}{Snapshot store directory.}
}
\description{
List snapshot ids in a store (newest last)
}
