% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pad_text}
\alias{pad_text}
\title{Pad text to a fixed width with a given alignment}
\usage{
pad_text(text, width, align = "left")
}
\description{
Pad text to a fixed width with a given alignment
}
\keyword{internal}
