% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_paragraph}
\alias{report_paragraph}
\title{A paragraph (markdown/plain text) section}
\usage{
report_paragraph(text)
}
\arguments{
\item{text}{Paragraph text (markdown is allowed).}
}
\value{
A \code{bctu_report_paragraph} section object.
}
\description{
A paragraph (markdown/plain text) section
}
