% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{run_pandoc}
\alias{run_pandoc}
\title{Run pandoc for one output format}
\usage{
run_pandoc(
  pandoc,
  md_path,
  out_path,
  fmt,
  template,
  toc = FALSE,
  number_sections = FALSE,
  lua_filter = NULL
)
}
\description{
Run pandoc for one output format
}
\keyword{internal}
