% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{save_snapshot}
\alias{save_snapshot}
\title{Save an in-memory snapshot to the store}
\usage{
save_snapshot(
  x,
  store = snapshot_location(),
  formats = c("rds", "csv"),
  verbose = 2L
)
}
\arguments{
\item{x}{A \code{bctu_snapshot}.}

\item{store}{Snapshot store directory.}

\item{formats}{Payload formats.}

\item{verbose}{Verbosity.}
}
\value{
\code{x} with \code{id} attached, invisibly.
}
\description{
Save an in-memory snapshot to the store
}
