% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sql_odbc_arguments}
\alias{sql_odbc_arguments}
\title{Build the ODBC \code{DBI::dbConnect()} argument list from an sql_connection}
\usage{
sql_odbc_arguments(server, database, conn)
}
\arguments{
\item{server}{SQL Server host.}

\item{database}{Database name.}

\item{conn}{An \code{\link[=sql_connection]{sql_connection()}}.}
}
\value{
A named list of connection arguments.
}
\description{
Build the ODBC \code{DBI::dbConnect()} argument list from an sql_connection
}
